Menambahkan Class Salary
Memindahkan atribut-atribut Salary kedalam class Salary
